improve little things, this commit is with unselected changes, because is failing, then I want to say that is the first time that I realiza a partial commit about all my changes
improve little things, this commit is with unselected changes, because is failing, then I want to say that is the first time that I realiza a partial commit about all my changes.

well, is the first that that I used correctly Github desktop, I explained me, how i comment before, I realized changes about Items to add a new camp, but it fail, well, is the first time that I delete changes since Github Desktop, and I love it because I recover the functionallity lost.